chore: Bump version to 0.21.0, update download links, and optimize SCSS with mixins and variables for improved maintainability
