+++ module/sheets/item-sheet.mjs
@@ -2,6 +2,9 @@
  * Extend the basic ItemSheet
  * @extends {ItemSheet}
  */
+import { prepareFiringModes, getLinkedDisciplineImage, enrichItemDescription } from "../helpers/item-sheet.mjs";
+import { handleWeaponDrop, handleWeaponSkillDrop, handleDisciplineDrop, handleSkillDrop, handleSkillDelete } from "../helpers/drop-handlers.mjs";
+
 export class SlaItemSheet extends foundry.appv1.sheets.ItemSheet {
 
     /** @override */
@@ -27,39 +30,16 @@ export class SlaItemSheet extends foundry.appv1.sheets.ItemSheet {
         context.flags = itemData.flags;
         context.config = CONFIG.SLA;
 
-        // --- 1. ENRICH DESCRIPTION (The Missing Part) ---
-        context.enrichedDescription = await foundry.applications.ux.TextEditor.enrichHTML(this.item.system.description, {
-            async: true,
-            relativeTo: this.item
-        });
+        // Enrich description
+        context.enrichedDescription = await enrichItemDescription(this.item);
 
-        // --- 2. PREPARE FIRING MODES (Data Model -> Object) ---
-        // Using toObject() ensures we work with POJOs. Explicit keys ensure iteration order.
+        // Prepare firing modes for weapons
         if (this.item.type === "weapon") {
-            context.firingModes = {};
-            const knownModes = ["single", "burst", "auto", "suppressive"];
-            const sourceModes = this.item.system.toObject().firingModes || {};
-
-            for (const key of knownModes) {
-                if (sourceModes[key]) {
-                    context.firingModes[key] = {
-                        ...sourceModes[key],
-                        id: key
-                    };
-                }
-            }
+            context.firingModes = prepareFiringModes(this.item.system);
         }
 
-        // --- 3. EBB IMAGE LOOKUP ---
-        if (this.item.actor && context.system.discipline) {
-            const disciplineItem = this.item.actor.items.find(i =>
-                i.type === "discipline" &&
-                i.name.toLowerCase() === context.system.discipline.toLowerCase()
-            );
-            context.linkedDisciplineImg = disciplineItem ? disciplineItem.img : "icons/svg/item-bag.svg";
-        } else {
-            context.linkedDisciplineImg = "icons/svg/item-bag.svg";
-        }
+        // Get linked discipline image for Ebb Formulas
+        context.linkedDisciplineImg = getLinkedDisciplineImage(this.item);
 
         return context;
     }
@@ -138,134 +118,28 @@ export class SlaItemSheet extends foundry.appv1.sheets.ItemSheet {
      * Handle dropping a Weapon item to link it to this Magazine
      */
     async _onDropWeapon(event) {
-        event.preventDefault();
-        try {
-            const data = JSON.parse(event.dataTransfer.getData('text/plain'));
-            if (data.type !== "Item") return;
-
-            const item = await Item.implementation.fromDropData(data);
-            if (item && item.type === "weapon") {
-                await this.item.update({ "system.linkedWeapon": item.name });
-                ui.notifications.info(`Linked Magazine to: ${item.name}`);
-            } else {
-                ui.notifications.warn("Only Weapons can be linked to a Magazine.");
-            }
-        } catch (err) {
-            console.error("SLA | Weapon Drop Failed:", err);
-        }
+        await handleWeaponDrop(event, this.item);
     }
 
     /**
      * Handle dropping a Skill item onto a Weapon to set requirement
      */
     async _onDropWeaponSkill(event) {
-        event.preventDefault();
-        try {
-            const data = JSON.parse(event.originalEvent.dataTransfer.getData('text/plain'));
-            if (data.type !== "Item") return;
-
-            const item = await Item.implementation.fromDropData(data);
-
-            // Validation: Must be a Skill
-            if (!item || item.type !== "skill") {
-                return ui.notifications.warn("Only 'Skill' items can be linked.");
-            }
-
-            // Save the NAME of the skill (e.g., "Pistol", "Melee")
-            await this.item.update({
-                "system.skill": item.name
-            });
-
-        } catch (err) {
-            console.error("SLA | Weapon Skill Drop Failed:", err);
-        }
+        await handleWeaponSkillDrop(event, this.item);
     }
 
     /**
      * Handle dropping a Discipline item onto an Ebb Formula
      */
     async _onDropDiscipline(event) {
-        event.preventDefault();
-        event.stopPropagation();
-
-        try {
-            const data = JSON.parse(event.dataTransfer.getData('text/plain'));
-            if (data.type !== "Item") return;
-
-            const item = await Item.implementation.fromDropData(data);
-
-            if (!item || item.type !== "discipline") {
-                return ui.notifications.warn("Only 'Discipline' items can be linked here.");
-            }
-
-            await this.item.update({
-                "system.discipline": item.name
-            });
-
-        } catch (err) {
-            console.error("SLA | Discipline Drop Failed:", err);
-        }
+        await handleDisciplineDrop(event, this.item);
     }
 
     /**
      * Handle dropping a Skill onto Species/Package (Creates List)
      */
     async _onDropSkill(event) {
-        event.preventDefault();
-        try {
-            // jQuery wraps the event, so we need originalEvent for dataTransfer
-            const data = JSON.parse(event.originalEvent.dataTransfer.getData('text/plain'));
-            if (data.type !== "Item") return;
-
-            const item = await Item.implementation.fromDropData(data);
-
-            // Validation
-            if (!item || item.type !== "skill") {
-                return ui.notifications.warn("Only Skills can be added to this list.");
-            }
-
-            // 1. Get the current list (safely)
-            let currentSkills = this.item.system.skills;
-            if (!Array.isArray(currentSkills)) {
-                currentSkills = [];
-            }
-
-            // MIGRATION: Fix any strings OR objects missing 'stat'
-            // If the user previously had an array of strings, we convert them now to avoid validation errors
-            const cleanSkills = currentSkills.map(s => {
-                if (typeof s === "string") {
-                    return {
-                        name: s,
-                        rank: 1,
-                        img: "icons/svg/item-bag.svg",
-                        stat: "dex"
-                    };
-                }
-                // Also ensure existing objects have stat
-                if (!s.stat) s.stat = "dex";
-                return s;
-            });
-
-            // 2. Build the data object
-            const newSkill = {
-                name: item.name,
-                rank: item.system.rank || 1,
-                img: item.img || "icons/svg/item-bag.svg",
-                stat: item.system.stat || "dex"
-            };
-
-            // 3. Check for duplicates
-            if (cleanSkills.some(s => s.name === newSkill.name)) {
-                return ui.notifications.warn(`${newSkill.name} is already in the list.`);
-            }
-
-            // 4. Update the Item
-            const newArray = [...cleanSkills, newSkill];
-            await this.item.update({ "system.skills": newArray });
-
-        } catch (err) {
-            console.error("SLA | Skill Drop Failed:", err);
-        }
+        await handleSkillDrop(event, this.item);
     }
 
     /**
@@ -273,19 +147,7 @@ export class SlaItemSheet extends foundry.appv1.sheets.ItemSheet {
      */
     async _onDeleteSkill(event) {
         event.preventDefault();
-        const target = event.currentTarget;
-        const index = Number(target.dataset.index);
-
-        const currentSkills = this.item.system.skills || [];
-
-        // Filter out the specific index AND sanitize remainder
-        const newArray = currentSkills
-            .filter((_, i) => i !== index)
-            .map(s => {
-                if (typeof s === "string") return { name: s, rank: 1, img: "icons/svg/item-bag.svg" };
-                return s;
-            });
-
-        await this.item.update({ "system.skills": newArray });
+        const index = Number(event.currentTarget.dataset.index);
+        await handleSkillDelete(index, this.item);
     }
 }
