+++ src/scss/components/_chat.scss
@@ -1,36 +1,42 @@
-/* ==========================================
-   10. CHAT CARD STYLING
-   ========================================== */
+/**
+ * Chat Card Styling
+ * Optimized with mixins and variables
+ */
+
+@import '../utils/mixins';
+@import '../utils/variables';
+
 .sla-chat-card {
-    font-family: "Josefin Sans", sans-serif;
+    @include font-body;
     font-size: 14px;
     color: #eee;
 
     h3 {
-        font-family: "Josefin Sans", sans-serif;
-        font-weight: bold;
-        text-transform: uppercase;
-        margin: 0 0 5px 0;
-        border-bottom: 1px solid #555;
+        @include font-header;
+        margin: 0 0 $padding-sm 0;
+        border-bottom: $border-width solid #555;
     }
 
     // New Class for Rerolled Dice
     .dice-rerolled {
-        border-style: dotted !important;
-        text-decoration-style: dotted !important;
+        border-style: dotted;
+        text-decoration-style: dotted;
     }
 
     // Difficulty Adjustment Buttons
     .sla-diff-btn {
-        background: #111;
-        border: 1px solid #444;
-        color: #aaa;
         padding: 2px 6px;
-        font-family: "Josefin Sans", sans-serif;
-        font-size: 0.9em;
+        border: $border-width solid #444;
+        border-radius: $border-radius-sm;
         cursor: pointer;
-        transition: all 0.2s ease;
-        border-radius: 2px;
+        font-family: var(--sla-font-mono);
+        font-size: 0.9em;
+        transition: all $transition-base;
+        text-transform: uppercase;
+        font-weight: bold;
+        background: #111;
+        border-color: #444;
+        color: #aaa;
 
         &:hover {
             background: #333;
@@ -38,26 +44,37 @@
             color: #fff;
             box-shadow: 0 0 5px rgba(255, 255, 255, 0.2);
         }
+        
+        &:focus {
+            outline: none;
+            box-shadow: 0 0 5px var(--sla-accent);
+        }
     }
 
     // Tactical Choice Buttons
     .mos-choice-container {
         button {
-            font-family: "Josefin Sans", sans-serif;
-            font-size: 0.9em;
             padding: 4px 8px;
+            border: $border-width solid;
+            border-radius: $border-radius-sm;
             cursor: pointer;
+            font-family: var(--sla-font-mono);
+            font-size: 0.9em;
+            transition: all $transition-base;
             text-transform: uppercase;
             font-weight: bold;
-            border-radius: 2px;
-            transition: all 0.2s ease;
+            
+            &:focus {
+                outline: none;
+                box-shadow: 0 0 5px var(--sla-accent);
+            }
         }
 
         .chat-btn-wound {
             background: #330000;
-            border: 1px solid #ff4444;
+            border-color: #ff4444;
             color: #ffcccc;
-
+            
             &:hover {
                 background: #550000;
                 box-shadow: 0 0 5px rgba(255, 68, 68, 0.4);
@@ -66,9 +83,9 @@
 
         .chat-btn-damage {
             background: #000;
-            border: 1px solid #39ff14;
+            border-color: #39ff14;
             color: #39ff14;
-
+            
             &:hover {
                 background: #002200;
                 box-shadow: 0 0 5px rgba(57, 255, 20, 0.4);
@@ -79,29 +96,24 @@
 
 /* Damage Result Card */
 .sla-chat-card.damage-result {
-    background: #000;
-    border: 1px solid #d00;
-    border-left: 4px solid #d00;
     padding: 8px;
-    border-radius: 4px;
+    border-radius: $border-radius;
+    background: #000;
+    border: $border-width solid #d00;
+    border-left: 4px solid #d00;
     box-shadow: 0 0 5px rgba(221, 0, 0, 0.2);
 
     .damage-header {
         color: #ff3333;
-        font-family: var(--sla-font-header);
-        font-weight: 700;
+        @include font-header;
         font-size: 1.1em;
-        text-transform: uppercase;
-        border-bottom: 1px solid #500;
+        border-bottom: $border-width solid #500;
         margin-bottom: 6px;
         padding-bottom: 2px;
-        letter-spacing: 0.5px;
     }
 
     .damage-row {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
+        @include flex-between;
         margin-bottom: 2px;
         font-size: 0.95em;
 
@@ -129,7 +141,7 @@
         text-align: right;
         margin-top: 8px;
         padding-top: 4px;
-        border-top: 1px solid #333;
+        border-top: $border-width solid #333;
         font-size: 0.9em;
         color: #bbb;
 
@@ -142,9 +154,8 @@
         font-size: 0.85em;
         color: #888;
         margin-top: 4px;
-        display: flex;
-        align-items: center;
-        gap: 5px;
+        @include flex-center;
+        gap: $padding-sm;
 
         i {
             color: #666;
@@ -154,23 +165,20 @@
 
 /* Drug Use Card */
 .sla-chat-card.drug-use {
-    background: #1a1a25;
-    border: 1px solid #d05e1a;
-    border-left: 4px solid #d05e1a;
     padding: 8px;
-    border-radius: 4px;
+    border-radius: $border-radius;
+    background: #000;
+    border: $border-width solid var(--sla-accent);
+    border-left: 4px solid var(--sla-accent);
     box-shadow: 0 0 5px rgba(208, 94, 26, 0.2);
 
     .drug-header {
-        color: #d05e1a;
-        font-family: var(--sla-font-header);
-        font-weight: 700;
+        color: var(--sla-accent);
+        @include font-header;
         font-size: 1.1em;
-        text-transform: uppercase;
-        border-bottom: 1px solid #555;
+        border-bottom: $border-width solid #555;
         margin-bottom: 6px;
         padding-bottom: 2px;
-        letter-spacing: 0.5px;
     }
 
     .drug-content {
