+++ src/scss/components/_dialog.scss
@@ -1,54 +1,64 @@
-/* ==========================================
-   11. DIALOG THEME (Attack & Luck)
-   ========================================== */
+/**
+ * Dialog Theme (Attack & Luck)
+ * Optimized with mixins and reduced !important usage
+ */
+
+@import '../utils/mixins';
+@import '../utils/variables';
+
 .sla-dialog-window.dialog {
-    background: url('') !important;
-    background-color: var(--sla-bg-dark) !important;
-    border: 2px solid var(--sla-border) !important;
+    background: url('');
+    background-color: var(--sla-bg-dark);
+    border: $border-width-thick solid var(--sla-border);
     box-shadow: 0 0 20px #000;
     color: var(--sla-text-light);
     border-radius: 5px;
 
     .window-header {
-        background: #111 !important;
-        border-bottom: 2px solid var(--sla-accent);
-        color: #fff !important;
-        font-family: var(--sla-font-header);
-        font-weight: 700;
-        text-transform: uppercase;
-        padding: 5px 10px;
+        background: #111;
+        border-bottom: $border-width-thick solid var(--sla-accent);
+        color: #fff;
+        @include font-header;
+        padding: $padding-sm $padding-md;
     }
 
     .window-content {
-        background: var(--sla-bg-dark) !important;
+        background: var(--sla-bg-dark);
         color: var(--sla-text-light);
-        padding: 10px;
+        padding: $padding-md;
         font-family: var(--sla-font-header);
     }
 
     select,
     input {
-        background-color: #000 !important;
-        color: #fff !important;
-        border: 1px solid var(--sla-border) !important;
+        background: #000;
+        color: var(--sla-text-light);
+        border: $border-width solid #555;
         border-radius: var(--sla-radius);
         height: 28px;
-        padding-left: 5px;
+        padding: 1px $padding-sm;
+        font-family: var(--sla-font-mono);
+        
+        &:focus {
+            box-shadow: 0 0 5px var(--sla-accent);
+            background: #111;
+            outline: none;
+        }
     }
 
     .dialog-buttons .dialog-button {
-        background: #2a2a35;
-        border: 1px solid var(--sla-accent);
+        background: var(--sla-bg-panel);
+        border: $border-width solid var(--sla-accent);
         color: var(--sla-accent);
-        font-family: var(--sla-font-header);
-        font-weight: 700;
-        text-transform: uppercase;
+        @include font-header;
         margin-top: 15px;
+        padding: $padding-sm $padding-md;
+        transition: all $transition-base;
 
         &:hover {
             background: var(--sla-accent);
             color: #000;
-            box-shadow: 0 0 15px var(--sla-accent-faint);
+            box-shadow: 0 0 10px var(--sla-accent);
         }
     }
 }
