+++ src/scss/components/_forms.scss
@@ -1,55 +1,63 @@
+/**
+ * Form Styles
+ * Optimized with mixins
+ */
+
+@import '../utils/mixins';
+@import '../utils/variables';
+
 .item-form {
-  font-family: $font-primary;
+    font-family: $font-primary;
 }
 
 .sheet-header {
-  flex: 0 auto;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  margin-bottom: 10px;
+    flex: 0 auto;
+    overflow: hidden;
+    @include flex-between;
+    flex-wrap: wrap;
+    justify-content: flex-start;
+    margin-bottom: $margin-md;
 
-  .profile-img {
-    flex: 0 0 100px;
-    height: 100px;
-    margin-right: 10px;
-  }
-
-  .header-fields {
-    flex: 1;
-  }
-
-  h1.charname {
-    height: 50px;
-    padding: 0px;
-    margin: 5px 0;
-    border-bottom: 0;
-    input {
-      width: 100%;
-      height: 100%;
-      margin: 0;
+    .profile-img {
+        flex: 0 0 100px;
+        height: 100px;
+        margin-right: $margin-md;
     }
-  }
+
+    .header-fields {
+        flex: 1;
+    }
+
+    h1.charname {
+        height: 50px;
+        padding: 0;
+        margin: $padding-sm 0;
+        border-bottom: 0;
+        
+        input {
+            width: 100%;
+            height: 100%;
+            margin: 0;
+        }
+    }
 }
 
 .sheet-tabs {
-  flex: 0;
+    flex: 0;
 }
 
 .sheet-body,
 .sheet-body .tab,
 .sheet-body .tab .editor {
-  height: 100%;
+    height: 100%;
 }
 
 .tox {
-  .tox-editor-container {
-    background: $c-white;
-  }
+    .tox-editor-container {
+        background: $c-white;
+    }
 
-  .tox-edit-area {
-    padding: 0 8px;
-  }
+    .tox-edit-area {
+        padding: 0 8px;
+    }
 }
