+++ src/scss/global/_base.scss
@@ -1,39 +1,43 @@
+/**
+ * Global Base Styles
+ * Uses mixins and variables for consistency
+ */
+
+@import '../utils/mixins';
+@import '../utils/variables';
+
 /* ==========================================
    0. FONTS
    ========================================== */
 @import url('https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@400;600;700&display=swap');
 
-
 /* ==========================================
-   1. GLOBAL VARIABLES & WINDOW STYLES
+   1. WINDOW STYLES
    ========================================== */
-/* WINDOW CONTENT (Dark Theme) */
 .sla-industries.sheet.actor .window-content,
 .sla-industries.sheet.item .window-content {
-    background-image: none !important;
-    background: var(--sla-bg-dark) !important;
-    background-color: var(--sla-bg-dark) !important;
+    background-image: none;
+    background: var(--sla-bg-dark);
+    background-color: var(--sla-bg-dark);
     color: var(--sla-text-light);
-    font-family: var(--sla-font-mono);
-    font-size: 13px;
-    padding: 0 !important;
-    overflow: hidden !important;
-    border: 2px solid var(--sla-border);
+    @include font-body;
+    padding: 0;
+    overflow: hidden;
+    border: $border-width-thick solid var(--sla-border);
 }
 
 /* FORM LAYOUT */
 form.sla-sheet {
     height: 100%;
-    display: flex;
-    flex-direction: column;
-    overflow: hidden !important;
-    background: transparent !important;
-    padding: 5px;
+    @include flex-column;
+    overflow: hidden;
+    background: transparent;
+    padding: $padding-sm;
 }
 
 /* UTILITIES */
 .sla-pad {
-    padding: 8px !important;
+    padding: 8px;
 }
 
 /* ==========================================
@@ -42,37 +46,36 @@ form.sla-sheet {
 .sla-sheet input[type="text"],
 .sla-sheet input[type="number"],
 .sla-sheet select {
-    background: #000;
-    color: #e0e0e0;
-    border: 1px solid #555;
-    border-radius: var(--sla-radius);
-    height: 26px;
-    padding: 1px 5px;
-    font-family: var(--sla-font-mono);
-}
-
-.sla-sheet input:focus,
-.sla-sheet select:focus {
-    box-shadow: 0 0 5px var(--sla-accent);
-    background: #111;
-    outline: none;
+    @include input-base;
 }
 
 /* ==========================================
-   12. PAUSE ANIMATION
+   3. DROPDOWN CONTRAST FIX
+   ========================================== */
+select {
+    background-color: #222;
+    color: #fff;
+    border: $border-width solid #444;
+
+    option {
+        background-color: #222;
+        color: #fff;
+    }
+}
+
+/* ==========================================
+   4. PAUSE ANIMATION
    ========================================== */
 #pause {
     background: rgba(0, 0, 0, 0.7);
     border-radius: 10px;
-    padding: 10px;
-    border: 1px solid var(--sla-border);
+    padding: $padding-md;
+    border: $border-width solid var(--sla-border);
 
     figcaption {
-        font-family: var(--sla-font-header);
-        font-weight: 700;
+        @include font-header;
         font-size: 2em;
         letter-spacing: 2px;
-        text-transform: uppercase;
         animation: slaRedPulse 2s infinite ease-in-out;
     }
 }
@@ -98,7 +101,6 @@ form.sla-sheet {
 }
 
 @keyframes slaIconPulse {
-
     0%,
     100% {
         filter: drop-shadow(0 0 0px #000);
@@ -108,19 +110,5 @@ form.sla-sheet {
     50% {
         filter: drop-shadow(0 0 10px rgba(255, 0, 0, 0.8));
         opacity: 1;
-    }
-}
-
-/* ==========================================
-   15. DROPDOWN CONTRAST FIX
-   ========================================== */
-select {
-    background-color: #222;
-    color: #fff;
-    border: 1px solid #444;
-
-    option {
-        background-color: #222;
-        color: #fff;
     }
 }
