+++ src/scss/global/_window.scss
@@ -1,12 +1,19 @@
+/**
+ * Window Styles
+ * Optimized with variables
+ */
+
+@import '../utils/variables';
+
 .window-app {
-  font-family: $font-primary;
+    font-family: $font-primary;
 }
 
 .rollable {
-  &:hover,
-  &:focus {
-    color: #000;
-    text-shadow: 0 0 10px red;
-    cursor: pointer;
-  }
+    &:hover,
+    &:focus {
+        color: #000;
+        text-shadow: 0 0 10px red;
+        cursor: pointer;
+    }
 }
